ai_engine/requirements.txt
langchain==0.1.0
langchain-core==0.1.10
langchain-community==0.0.13
langgraph==0.0.20
google-genai==0.3.0
chromadb==0.4.22
PyPDF2==3.0.1
python-docx==1.1.0
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
aiohttp==3.9.1
asyncio==3.4.3
python-json-logger==2.0.7
black==23.12.1
flake8==7.0.0
mypy==1.8.0